Move HostInfo to its own package
HostInfo is not only used by sciond but also by sig mgmt therefore we move it to its own package.
